Fix clippy warnings across workspace
Resolve all clippy lints reported by `cargo clippy --workspace
--all-targets --all-features` (cargo fmt reported no changes):

- clarirs_core: remove identity `.into()` conversions on `AstRef`
  (useless_conversion) in find_variable and post_order tests; drop
  trailing underscores from hex literals (unusual_byte_groupings) in
  the extract test.
- clarirs_num: drop redundant `u64::MAX as u64` cast (unnecessary_cast).
- clarirs_z3: remove needless borrow of a generic arg and use
  `std::slice::from_ref` instead of cloning into a one-element slice in
  the solver tests; replace PI/E float literals with the corresponding
  `std::{f32,f64}::consts` constants (approximate_constant) in the
  float round-trip tests.

+++ crates/clarirs_num/src/bitvec/conversion.rs
@@ -299,7 +299,7 @@ mod tests {
         assert_eq!(bv.len(), 64);
         assert_eq!(bv.to_u64().unwrap(), 0);
 
-        let bv = BitVec::from((u64::MAX as u64, 64));
+        let bv = BitVec::from((u64::MAX, 64));
         assert_eq!(bv.len(), 64);
         assert_eq!(bv.to_u64().unwrap(), u64::MAX);
     }
